API-ified. All CoreData types (not config currently) are using API.TYPE.Action
