+++ Plugins/CoreAPI/server/graphql/resolvers/service.default.js
@@ -1,12 +1,13 @@
 import PenPal from "meteor/penpal";
-import { Mongo } from "meteor/mongo";
 
 export default {
   Service: {
     host(service) {
-      return PenPal.DataStore.fetch("CoreAPI", "Hosts", {
-        _id: new Mongo.ObjectID(service.hostID),
-      })[0];
+      return PenPal.API.Hosts.Get({ id: service.hostID });
+    },
+    project(service) {
+      console.log(service.projectID);
+      return PenPal.API.Projects.Get({ id: service.projectID });
     },
   },
 };
